+++ R/data_input.R
@@ -130,7 +130,7 @@ read_expression_data = function(expr_list = NULL,
 
 
   # 1. get depth of list
-  # if(verbose == TRUE) print(str(expr_list))
+
   list_depth = depth(expr_list)
 
 
@@ -677,7 +677,7 @@ read_spatial_location_data = function(spat_loc_list,
 
 
   # 1. get depth of list
-  # if(verbose == TRUE) print(str(spat_loc_list))
+
   list_depth = depth(spat_loc_list)
 
   # no expression information
@@ -1919,7 +1919,7 @@ extract_polygon_list = function(polygonlist,
 
   }
 
-  # print(final_list)
+
   return(final_list)
 
 }